Tic Tac Toe algorithm with preferences
This is a Tic Tac Toe algorithm that plays as O and the user plays as X. It has preferences that someone playing with O would usually have. It first tries to find a winning move, if it cannot do so, it looks if the opponent can win in the next move and tries to stop it. If none of the above is true it tries to play in the center. And if that is also not true it tries to play adjacent to a previously placed O to get a chance of winning and forcing the opponent to play some of the moves that he wouldn't play otherwise.